Move logic of checking available port to service
